test: Add unit tests for extension functions
This commit adds unit tests for the existing extension functions for `Boolean`, `Double`, `Int`, `Long`, and `String`.

Key changes:
- Created `BooleanExtensionsTest.kt` with tests for `orFalse` and `orTrue`.
- Created `DoubleExtensionsTest.kt` with tests for `defaultValue`, `isZero`, `orZero`, and `toStringSafe`.
- Created `IntExtensionsTest.kt` with tests for `defaultValue`, `isZero`, `orZero`, and `toStringSafe`.
- Created `LongExtensionsTest.kt` with tests for `defaultValue`, `isZero`, `orZero`, and `toStringSafe`.
- Created `StringExtensionsTest.kt` with tests for `EMPTY` and `orDefault`.

Additionally, a bug in `LongExtensions.kt` was fixed where `orZero()` was incorrectly returning `Int.ZERO` instead of `Long.ZERO`.

+++ app/src/main/java/br/com/seucaio/cryptoexchanges/core/extension/LongExtensions.kt
@@ -6,6 +6,6 @@ val Long.Companion.ZERO get() = 0L
 
 fun Long.isZero() = this == Long.ZERO
 
-fun Long?.orZero() = this ?: Int.ZERO
+fun Long?.orZero() = this ?: Long.ZERO
 
 fun Long?.toStringSafe() : String = defaultValue().toString()
